src/data_pipeline/pyproject.toml
[tool.poetry]
name = "data_pipeline"
version = "0.1.0"
description = ""
authors = ["Felipe Saadi <felipesaadisiegert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
clickhouse-connect = "^0.7.19"
python-dotenv = "^1.0.1"
Flask = "^3.0.3"
requests = "^2.32.3"
supabase = "^2.7.3"


[tool.poetry.dev-dependencies]
pytest = "^8.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
